+++ CHANGELOG.md
@@ -40,6 +40,7 @@
 * Add `SecretCreateOptsBuilder::labels`
 * Add `SystemdUnitsOptsBuilder::additional_env_variables`
 * Add `SystemdUnitsOptsBuilder::after`
+* Add `SystemdUnitsOptsBuilder::requires`
 
 # 0.7.0
 * Make futures returned by methods Send + Sync